requirements.txt
Office365-REST-Python-Client
pandas
customtkinter
Pillow
paramiko